+++ src/main/main.ts
@@ -5,8 +5,8 @@
  */
 import path from 'path';
 import { app, BrowserWindow, shell, ipcMain, dialog, Tray, Menu } from 'electron';
-import { resolveHtmlPath, loadAllVideos, isConfigReady, deleteVideo, openSystemExplorer, toggleVideoProtected, fixPathWhenPackaged, getPathConfigSafe, getNumberConfigSafe, defaultMonitorIndex, defaultMinEncounterDuration, getStringConfigSafe } from './util';
-import { watchLogs, pollWowProcess, runRecordingTest } from './logutils';
+import { resolveHtmlPath, loadAllVideos, isConfigReady, deleteVideo, openSystemExplorer, toggleVideoProtected, fixPathWhenPackaged, getPathConfigSafe, getNumberConfigSafe, defaultMonitorIndex, defaultMinEncounterDuration, getStringConfigSafe, defaultAudioDevice } from './util';
+import { watchLogs, pollWowProcess, runRecordingTest, forceStopRecording } from './logutils';
 import Store from 'electron-store';
 const obsRecorder = require('./obsRecorder');
 import { Recorder, RecorderOptionsType } from './recorder';
@@ -64,6 +64,14 @@ const loadRecorderOptions = (cfg: ElectronStore): RecorderOptionsType => {
 
   if (!config.minEncounterDuration) {
     config.minEncounterDuration = defaultMinEncounterDuration(cfg);
+  }
+
+  if (!config.audioInputDeviceId) {
+    config.audioInputDeviceId = defaultAudioDevice(cfg, 'input');
+  }
+  
+  if (!config.audioOutputDeviceId) {
+    config.audioOutputDeviceId = defaultAudioDevice(cfg, 'output');
   }
 
   return config;
@@ -494,12 +502,22 @@ ipcMain.on('videoPlayerSettings', (event, args) => {
 /**
  * Test button listener. 
  */
-ipcMain.on('test', () => {
+ipcMain.on('test', (_event, args) => {
   if (isConfigReady(cfg)) { 
     console.info("[Main] Config is good, running test!");
-    runRecordingTest()
+    runRecordingTest(Boolean(args[0]));
   } else {
     console.info("[Main] Config is bad, don't run test");
+  }
+});
+
+/**
+ * Recorder IPC functions
+ */
+ipcMain.on('recorder', (_event, args) => {
+  if (args[0] == 'stop') {
+    console.log('[Main] Force stopping recording due to user request.')
+    forceStopRecording();
   }
 });