+++ src/renderer/TestButton.tsx
@@ -4,8 +4,12 @@ const ipc = window.electron.ipcRenderer;
 
 export default function TestButton() {
 
-  const runTest = () => {
-    ipc.sendMessage('test', []);
+  const runTest = (event: any) => {
+    // 'Click' will perform a normal test
+    // 'Ctrl-Alt-Click' will initiate a test but won't finish it
+    // and requires a force stop of the recording.
+    const endTest = !(event.ctrlKey && event.altKey);
+    ipc.sendMessage('test', [endTest]);
   };
 
   return (